introduce QueryDialect parent to SqlDialect
